Handling multiple buffers and mixing them in SA

+++ src/Engine/Audio/SoundBuffer.cpp
@@ -20,7 +20,7 @@ SoundBuffer::~SoundBuffer() {
     free(buffer);
 }
 
-short* SoundBuffer::GetBuffer() const {
+float* SoundBuffer::GetBuffer() const {
     return buffer;
 }
